Add BaseDevice and BtHomeV2Device classes for BTHome V2 integration
- Introduced BaseDevice class to handle common device functionalities.
- Created BtHomeV2Device class for specific sensor data handling.
- Implemented methods for adding various sensor data types including temperature, distance, humidity, and more.
- Defined data types and states for BTHome V2 protocol in data_types.h.
- Added definitions for constants and flags in definitions.h.
- Updated versioning in _version.py and enhanced logging capabilities in bthome_logger.py.

+++ examples/ESP32_MultipleSensors/src/main.cpp
@@ -87,7 +87,7 @@ void updateSensors() {
   humidity += (random(-100, 100) / 100.0);
   pressure += (random(-50, 50) / 100.0);
   co2 += random(-10, 10);
-  battery = max(0, battery - (random(0, 2)));  // Slow battery drain
+  battery = max(0, (int)(battery - random(0, 2)));  // Slow battery drain
 
   // Keep values in reasonable ranges
   temperature = constrain(temperature, -20.0, 40.0);
@@ -104,17 +104,14 @@ void updateSensors() {
   Serial.printf("Motion:       %s\n", motionDetected ? "YES" : "NO");
   Serial.printf("Door:         %s\n", doorOpen ? "OPEN" : "CLOSED");
 
-  // Clear old measurements
+  // Clear old measurements and add new ones
   bthome.clearMeasurements();
 
-  // Add all measurements
   bthome.addTemperature(temperature);
   bthome.addHumidity(humidity);
   bthome.addPressure(pressure);
   bthome.addCO2(co2);
   bthome.addBattery(battery);
-
-  // Add binary sensors
   bthome.addBinarySensor(MOTION, motionDetected);
   bthome.addBinarySensor(DOOR, doorOpen);
